Format 'classer.py' and change 'threshold' parser for 'samples'
